move test into separate file
